move and update validator
